+++ spec/fixtures/index.js
@@ -1,1 +1,2 @@
+require('./common');
 document.body.innerHTML = document.body.innerHTML + "<p>index.js</p>";